Refactored rag folder into services
